refactor(eu): extract shared GspaApiEU base
Login CCI + token refresh + stamp + GSPA machinery moved to
GspaApiEU(ApiImpl). HyundaiCciApiEU now inherits, keeping only
Hyundai-specific reads. Device-header is a per-brand class
attribute (§3.1). Test patch targets migrated to base.
